Change how job exit state us handled in the job runner.

+++ cli/process_jobs.go
@@ -21,14 +21,217 @@ import (
 )
 
 var exitOnFailure = false
+var runnerDir string
+var actuallyPowerThings = true
 
-func Log(uuid *strfmt.UUID, s string) error {
-	buf := bytes.NewBufferString(s)
+var cmdHelper = []byte(`
+#!/bin/bash
+
+# To force dpkg on Debian-based distros to play nice.
+export DEBIAN_FRONTEND=noninteractive DEBCONF_NONINTERACTIVE_SEEN=true
+
+# Force everything to use the C locale to keep things sane
+export LC_ALL=C LANGUAGE=C LANG=C
+
+# Make sure we play nice with debugging
+export PS4='${BASH_SOURCE}@${LINENO}(${FUNCNAME[0]}): '
+
+# Make sure the scripts are somewhat typo-resistant
+set -o pipefail -o errexit
+shopt -s nullglob extglob globstar
+
+# Make sure that $PATH is somewhat sane.
+fix_path() {
+    local -A pathparts
+    local part
+    local IFS=':'
+    for part in $PATH; do
+        pathparts["$part"]="true"
+    done
+    local wanted_pathparts=("/usr/local/bin" "/usr/local/sbin" "/bin" "/sbin" "/usr/bin" "/usr/sbin")
+    for part in "${wanted_pathparts[@]}"; do
+        [[ ${pathparts[$part]} ]] && continue
+        PATH="$part:$PATH"
+    done
+}
+fix_path
+unset fix_path
+
+# Figure out what Linux distro we are running on.
+export OS_TYPE= OS_VER= OS_NAME=
+if [[ -f /etc/os-release ]]; then
+    . /etc/os-release
+    OS_TYPE=${ID,,}
+    OS_VER=${VERSION_ID,,}
+elif [[ -f /etc/lsb-release ]]; then
+    . /etc/lsb-release
+    OS_VER=${DISTRIB_RELEASE,,}
+    OS_TYPE=${DISTRIB_ID,,}
+elif [[ -f /etc/centos-release || -f /etc/fedora-release || -f /etc/redhat-release ]]; then
+    for rel in centos-release fedora-release redhat-release; do
+        [[ -f /etc/$rel ]] || continue
+        OS_TYPE=${rel%%-*}
+        OS_VER="$(egrep -o '[0-9.]+' "/etc/$rel")"
+        break
+    done
+    if [[ ! $OS_TYPE ]]; then
+        echo "Cannot determine Linux version we are running on!"
+        exit 1
+    fi
+elif [[ -f /etc/debian_version ]]; then
+    OS_TYPE=debian
+    OS_VER=$(cat /etc/debian_version)
+fi
+OS_NAME="$OS_TYPE-$OS_VER"
+
+case $OS_TYPE in
+    centos|redhat|fedora|rhel|scientificlinux) OS_FAMILY="rhel";;
+    debian|ubuntu) OS_FAMILY="debian";;
+    *) OS_FAMILY=$OS_TYPE;;
+esac
+
+if_update_needed() {
+    local timestampref=/tmp/pkg_cache_update
+    if [[ ! -f $timestampref ]] || \
+           (( ($(stat -c '%Y' "$timestampref") - $(date '+%s')) > 86400 )); then
+        touch "$timestampref"
+        "$@"
+    fi
+}
+
+# Install a package
+install() {
+    local to_install=()
+    local pkg
+    for pkg in "$@"; do
+        to_install+=("$pkg")
+    done
+    case $OS_FAMILY in
+        rhel)
+            if_update_needed yum -y makecache
+            yum -y install "${to_install[@]}";;
+        debian)
+            if_update_needed apt-get -y update
+            apt-get -y install "${to_install[@]}";;
+        alpine)
+            if_update_needed apk update
+            apk add "${to_install[@]}";;
+        *) echo "No idea how to install packages on $OS_NAME"
+           exit 1;;
+    esac
+}
+
+INITSTYLE="sysv"
+if which systemctl &>/dev/null; then
+    INITSTYLE="systemd"
+elif which initctl &>/dev/null; then
+    INITSTYLE="upstart"
+fi
+
+# Perform service actions.
+service() {
+    # $1 = service name
+    # $2 = action to perform
+    local svc="$1"
+    shift
+    if which systemctl &>/dev/null; then
+        systemctl "$1" "$svc.service"
+    elif which chkconfig &>/dev/null; then
+        case $1 in
+            enable) chkconfig "$svc" on;;
+            disable) chkconfig "$svc" off;;
+            *)  command service "$svc" "$@";;
+        esac
+    elif which initctl &>/dev/null && initctl version 2>/dev/null | grep -q upstart ; then
+        /usr/sbin/service "$svc" "$1"
+    elif [[ -f /etc/init/$svc.unit ]]; then
+        initctl "$1" "$svc"
+    elif which update-rc.d &>/dev/null; then
+        case $1 in
+            enable|disable) update-rc.d "$svc" "$1";;
+            *) "/etc/init.d/$svc" "$1";;
+        esac
+    elif [[ -x /etc/init.d/$svc ]]; then
+        "/etc/init.d/$svc" "$1"
+    else
+        echo "No idea how to manage services on $OS_NAME"
+        exit 1
+    fi
+}
+
+get_param() {
+    # $1 attrib to get.  Attrib will be fetched in the context of the current machine
+    local attr
+    drpcli machines get "$RS_UUID" param "$1"
+}
+
+set_param() {
+    # $1 = name of the parameter to set
+    # $2 = parameter to set.
+    #      if $2 == "", then we will read from stdin
+    local src="$2"
+    if [[ ! $src ]]; then src="-"; fi
+    drpcli machines set "$RS_UUID" param "$1" to "$src"
+}
+
+__sane_exit() {
+    touch "$RS_TASK_DIR/.sane-exit-codes"
+}
+
+__exit() {
+    __sane_exit
+    exit $1
+}
+
+exit_incomplete() {
+    __exit 128
+}
+
+exit_reboot() {
+    __exit 64
+}
+
+exit_shutdown() {
+    __exit 32
+}
+
+exit_incomplete_reboot() {
+    __exit 192
+}
+
+exit_incomplete_shutdown() {
+    __exit 160
+}
+
+addr_port() {
+    if [[ $1 =~ ':' ]]; then
+        printf '[%s]:%d' "$1" "$2"
+    else
+        printf '%s:%d' "$1" "$2"
+    fi
+}
+
+if ! (which jq &>/dev/null || install jq); then
+    echo "JQ not installed and not installable.  The script jig requires it to function"
+    exit 1
+fi
+`)
+
+func putLog(uuid *strfmt.UUID, buf *bytes.Buffer) error {
 	_, err := session.Jobs.PutJobLog(jobs.NewPutJobLogParams().WithUUID(*uuid).WithBody(buf), basicAuth)
 	if err != nil {
 		fmt.Printf("Failed to log to job log, %s: %v\n", uuid.String(), err)
 	}
 	return err
+}
+
+func Log(uuid *strfmt.UUID, echo bool, s string, args ...interface{}) error {
+	buf := &bytes.Buffer{}
+	fmt.Fprintf(buf, s, args...)
+	if echo {
+		fmt.Printf(s, args...)
+	}
+	return putLog(uuid, buf)
 }
 
 func writeStringToFile(filename, content string) error {
@@ -74,6 +277,7 @@ type CommandRunner struct {
 	stderr   io.ReadCloser
 	stdout   io.ReadCloser
 	stdin    io.WriteCloser
+	task     *models.Task
 	finished chan bool
 }
 
@@ -81,7 +285,7 @@ func (cr *CommandRunner) ReadLog() {
 	// read command's stderr line by line - for logging
 	in := bufio.NewScanner(cr.stderr)
 	for in.Scan() {
-		Log(cr.uuid, in.Text())
+		putLog(cr.uuid, bytes.NewBuffer(in.Bytes()))
 	}
 	cr.finished <- true
 }
@@ -90,20 +294,18 @@ func (cr *CommandRunner) ReadReply() {
 	// read command's stdout line by line - for replies
 	in := bufio.NewScanner(cr.stdout)
 	for in.Scan() {
-		Log(cr.uuid, in.Text())
+		putLog(cr.uuid, bytes.NewBuffer(in.Bytes()))
 	}
 	cr.finished <- true
 }
 
-func (cr *CommandRunner) Run() (failed, incomplete, reboot bool) {
+func (cr *CommandRunner) Run() (failed, incomplete, reboot, poweroff bool) {
 	// Start command running
 	err := cr.cmd.Start()
 	if err != nil {
 		failed = true
 		reboot = false
-		s := fmt.Sprintf("Command %s failed to start: %v\n", cr.name, err)
-		fmt.Printf(s)
-		Log(cr.uuid, s)
+		Log(cr.uuid, true, "Command %s failed to start: %v\n", cr.name, err)
 		return
 	}
 
@@ -119,82 +321,92 @@ func (cr *CommandRunner) Run() (failed, incomplete, reboot bool) {
 		// syscall is generally platform dependent, WaitStatus is
 		// defined for both Unix and Windows and in both cases has
 		// an ExitStatus() method with the same signature.
-		if status, ok := exiterr.Sys().(syscall.WaitStatus); ok {
-			code := status.ExitStatus()
-			switch code {
-			case 0:
-				failed = false
-				reboot = false
-				s := fmt.Sprintf("Command %s succeeded\n", cr.name)
-				fmt.Printf(s)
-				Log(cr.uuid, s)
-			case 1:
-				failed = false
-				reboot = true
-				s := fmt.Sprintf("Command %s succeeded (wants reboot)\n", cr.name)
-				fmt.Printf(s)
-				Log(cr.uuid, s)
-			case 2:
-				incomplete = true
-				reboot = false
-				s := fmt.Sprintf("Command %s incomplete\n", cr.name)
-				fmt.Printf(s)
-				Log(cr.uuid, s)
-			case 3:
-				incomplete = true
-				reboot = true
-				s := fmt.Sprintf("Command %s incomplete (wants reboot)\n", cr.name)
-				fmt.Printf(s)
-				Log(cr.uuid, s)
-			default:
+		status, statusOK := exiterr.Sys().(syscall.WaitStatus)
+		if !statusOK {
+			if err != nil {
 				failed = true
 				reboot = false
-				s := fmt.Sprintf("Command %s failed\n", cr.name)
-				fmt.Printf(s)
-				Log(cr.uuid, s)
+			} else {
+				failed = false
+				reboot = false
+			}
+		} else {
+			sane := false
+			if cr.task.Meta != nil {
+				flagStr, ok := cr.task.Meta["feature-flags"]
+				if ok {
+					for _, testFlag := range strings.Split(flagStr, ",") {
+						if "sane-exit-codes" == strings.TrimSpace(testFlag) {
+							sane = true
+							break
+						}
+					}
+				}
+			}
+			if !sane {
+				if st, err := os.Stat(".sane-exit-codes"); err == nil && st.Mode().IsRegular() {
+					sane = true
+				}
+			}
+			code := uint(status.ExitStatus())
+			if sane {
+				// codes can be between 0 and 255
+				// if the low bits are not 0, the command failed.
+				failed = code&^224 > uint(0)
+				// If the high bit is set, the command was incomplete and the
+				// the current task pointer should not be advanced.
+				incomplete = code&128 > uint(0)
+				// If we need a reboot, set bit 6
+				reboot = code&64 > uint(0)
+				// If we need to poweroff, set bit 5.  Reboot wins if it is set.
+				poweroff = code&32 > uint(0)
+			} else {
+				switch code {
+				case 0:
+				case 1:
+					reboot = true
+				case 2:
+					incomplete = true
+				case 3:
+					incomplete = true
+					reboot = true
+				default:
+					failed = true
+				}
 			}
 		}
 	} else {
 		if err != nil {
 			failed = true
 			reboot = false
-			s := fmt.Sprintf("Command %s failed: %v\n", cr.name, err)
-			fmt.Printf(s)
-			Log(cr.uuid, s)
 		} else {
 			failed = false
 			reboot = false
-			s := fmt.Sprintf("Command %s succeeded\n", cr.name)
-			fmt.Printf(s)
-			Log(cr.uuid, s)
 		}
 	}
-
+	Log(cr.uuid, true, "Command %s: failed: %v, incomplete: %v, reboot: %v, poweroff: %v\n",
+		cr.name,
+		failed,
+		incomplete,
+		reboot,
+		poweroff)
 	// Remove script
 	os.Remove(cr.cmd.Path)
 
 	return
 }
 
-func NewCommandRunner(uuid *strfmt.UUID, name, content string) (*CommandRunner, error) {
+func NewCommandRunner(uuid *strfmt.UUID, name, content, loc string) (*CommandRunner, error) {
 	answer := &CommandRunner{name: name, uuid: uuid}
-
-	// Make script file
-	tmpFile, err := ioutil.TempFile(".", "script")
-	if err != nil {
+	if err := ioutil.WriteFile(path.Join(loc, "script"), []byte(content), 0700); err != nil {
 		return nil, err
 	}
-	if _, err := tmpFile.Write([]byte(content)); err != nil {
+	if err := ioutil.WriteFile(path.Join(loc, "helper"), cmdHelper, 0400); err != nil {
 		return nil, err
 	}
-	path := "./" + tmpFile.Name()
-	if err := tmpFile.Close(); err != nil {
-		return nil, err
-	}
-	os.Chmod(path, 0700)
-
-	answer.cmd = exec.Command(path)
-
+	answer.cmd = exec.Command("./script")
+	answer.cmd.Dir = loc
+	answer.cmd.Env = append(os.Environ(), "RS_TASK_DIR="+loc, "RS_RUNNER_DIR="+runnerDir)
 	var err2 error
 	answer.stderr, err2 = answer.cmd.StderrPipe()
 	if err2 != nil {
@@ -216,33 +428,37 @@ func NewCommandRunner(uuid *strfmt.UUID, name, content string) (*CommandRunner, 
 	return answer, nil
 }
 
-func runContent(uuid *strfmt.UUID, action *models.JobAction) (failed, incomplete, reboot bool) {
+func runContent(uuid *strfmt.UUID, action *models.JobAction, task *models.Task) (failed, incomplete, reboot, poweroff bool) {
+	tmpDir, err := ioutil.TempDir(runnerDir, *task.Name+"-")
+	if err != nil {
+		Log(uuid, true, "Could not create temp dir: %v", err)
+	}
+	defer os.RemoveAll(tmpDir)
 
-	Log(uuid, fmt.Sprintf("Starting Content Execution for: %s\n", *action.Name))
+	Log(uuid, false, "Starting Content Execution for: %s\n", *action.Name)
 
-	runner, err := NewCommandRunner(uuid, *action.Name, *action.Content)
+	runner, err := NewCommandRunner(uuid, *action.Name, *action.Content, tmpDir)
 	if err != nil {
 		failed = true
-		s := fmt.Sprintf("Creating command %s failed: %v\n", *action.Name, err)
-		fmt.Printf(s)
-		Log(uuid, s)
-	} else {
-		failed, incomplete, reboot = runner.Run()
+		Log(uuid, true, "Creating command %s failed: %v\n", *action.Name, err)
+		return
 	}
-	return
+	runner.task = task
+	return runner.Run()
 }
 
 func processJobsCommand() *cobra.Command {
 	mo := &MachineOps{CommonOps{Name: "machines", SingularName: "machine"}}
 	jo := &JobOps{CommonOps{Name: "jobs", SingularName: "job"}}
 	so := &StageOps{CommonOps{Name: "stages", SingularName: "stage"}}
+	to := &TaskOps{CommonOps{Name: "tasks", SingularName: "task"}}
 
 	command := &cobra.Command{
 		Use:   "processjobs [id]",
 		Short: "For the given machine, process pending jobs until done.",
 		Long: `
 For the provided machine, identified by UUID, process the task list on
-that machine until an error occurs or all jobs are complete.  Upon 
+that machine until an error occurs or all jobs are complete.  Upon
 completion, optionally wait for additional jobs as specified by
 the stage runner wait flag.
 `,
@@ -263,6 +479,12 @@ the stage runner wait flag.
 			} else {
 				machine = obj.(*models.Machine)
 			}
+			var err error
+			runnerDir, err = ioutil.TempDir("", "runner-")
+			if err != nil {
+				return generateError(err, "Error making temp runner dir")
+			}
+			defer os.RemoveAll(runnerDir)
 
 			fmt.Printf("Processing jobs for %s\n", uuid)
 
@@ -328,13 +550,19 @@ the stage runner wait flag.
 					job = obj.(*models.Job)
 				}
 				did_job = true
+				var task *models.Task
+
+				if obj, err := Get(job.Task, to); err != nil {
+					Log(job.UUID, true, "Error loading task content: %v, continuing", err)
+					markJob(job.UUID.String(), "failed", jo)
+				} else {
+					task = obj.(*models.Task)
+				}
 
 				// Get the job data
 				var list []*models.JobAction
 				if resp, err := session.Jobs.GetJobActions(jobs.NewGetJobActionsParams().WithUUID(*job.UUID), basicAuth); err != nil {
-					s := fmt.Sprintf("Error loading task content: %v, continuing", err)
-					fmt.Printf(s)
-					Log(job.UUID, s)
+					Log(job.UUID, true, "Error loading task content: %v, continuing", err)
 					markJob(job.UUID.String(), "failed", jo)
 					continue
 				} else {
@@ -343,9 +571,7 @@ the stage runner wait flag.
 
 				// Mark job as running
 				if _, err := Update(job.UUID.String(), `{"State": "running"}`, jo, false); err != nil {
-					s := fmt.Sprintf("Error marking job as running: %v, continue\n", err)
-					fmt.Printf(s)
-					Log(job.UUID, s)
+					Log(job.UUID, true, "Error marking job as running: %v, continue\n", err)
 					markJob(job.UUID.String(), "failed", jo)
 					continue
 				}
@@ -354,6 +580,7 @@ the stage runner wait flag.
 				failed := false
 				incomplete := false
 				reboot := false
+				poweroff := false
 				state := "finished"
 
 				for _, action := range list {
@@ -366,18 +593,15 @@ the stage runner wait flag.
 					// Excute task
 					if *action.Path == "" {
 						fmt.Printf("Running Task Template: %s\n", *action.Name)
-						failed, incomplete, reboot = runContent(job.UUID, action)
+						failed, incomplete, reboot, poweroff = runContent(job.UUID, action, task)
 					} else {
 						fmt.Printf("Putting Content in place for Task Template: %s\n", *action.Name)
-						var s string
 						if err := writeStringToFile(*action.Path, *action.Content); err != nil {
 							failed = true
-							s = fmt.Sprintf("Task Template: %s - Copying contents to %s failed\n%v", *action.Name, *action.Path, err)
+							Log(job.UUID, true, "Task Template: %s - Copying contents to %s failed\n%v", *action.Name, *action.Path, err)
 						} else {
-							s = fmt.Sprintf("Task Template: %s - Copied contents to %s successfully\n", *action.Name, *action.Path)
+							Log(job.UUID, true, "Task Template: %s - Copied contents to %s successfully\n", *action.Name, *action.Path)
 						}
-						fmt.Printf(s)
-						Log(job.UUID, s)
 					}
 
 					if failed {
@@ -393,7 +617,7 @@ the stage runner wait flag.
 						fmt.Printf("Error posting event: %v\n", err)
 					}
 
-					if failed || incomplete || reboot {
+					if failed || incomplete || reboot || poweroff {
 						break
 					}
 				}
@@ -403,9 +627,23 @@ the stage runner wait flag.
 				// Loop back and wait for the machine to get marked runnable again
 
 				if reboot {
-					_, err := exec.Command("reboot").Output()
-					if err != nil {
-						Log(job.UUID, "Failed to issue reboot\n")
+					if actuallyPowerThings {
+						_, err := exec.Command("reboot").Output()
+						if err != nil {
+							Log(job.UUID, false, "Failed to issue reboot\n")
+						}
+					} else {
+						Log(job.UUID, true, "Would have rebooted")
+					}
+				}
+				if poweroff {
+					if actuallyPowerThings {
+						_, err := exec.Command("poweroff").Output()
+						if err != nil {
+							Log(job.UUID, false, "Failed to issue poweroff\n")
+						}
+					} else {
+						Log(job.UUID, true, "Would have powered down")
 					}
 				}
